// src/app/modules/academics/student-information-system/model/student-group-details-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StudentGroupDetailsView
 */
export interface StudentGroupDetailsView {
    sortBy: string;
    sortOrder?: number;
    shortNames?: Array<string>;
    fullNames?: Array<string>;
    symbols?: Array<string>;
    Motos?: Array<string>;
    flags?: Array<string>;
    pageNumber: number;
    pageSize: number;
}