// src/app/modules/academics/student-information-system/model/assign-group-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttachmentViewModel } from './attachment-view-model';
import { AssignRepresentativeView } from './assign-representative-view';


/**
 * SchoolRepRoleView
 */
export interface AssignGroupView { 
    /**
     * Gets or sets
     */
    studentGroupId?: string;
    /**
     * Gets or sets  ShortName
     */
    shortName?: string;
    /**
     * Gets or sets  ShortName
     */
    fullName?: string;
    /**
     * Gets or sets  Color
     */
    color?: string;
    /**
     * Gets or sets  SymbolName
     */
    symbolName?: string;
    /**
     * Gets or sets  Moto
     */
    moto?: string;
    /**
     * Gets or sets  Moto
     */
    flag?: string;
    /**
     * Gets or sets  FlagAttachmentId
     */
    flagAttachmentId?: number;
    /**
     * Gets or sets Attachment
     */
    attachment?: AttachmentViewModel;
    /**
     * Gets or sets  GroupRepresentativesView
     */
    studentGroupReps?: Array<AssignRepresentativeView>;
}