// src/app/modules/academics/student-information-system/model/assign-representative-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * GroupRepresentativesView
 */
export interface AssignRepresentativeView {
    /**
     * Gets or sets  studentGroupRepRoleId
     */
    studentGroupRepRoleId?: number;
    /**
     * Gets or sets  studentId
     */
    studentId?: number;
    /**
     * Gets or sets  staffId
     */
    staffId?: number;

}